Add a common.py file and export some classes
